+++ app/Models/NetworkHost.php
@@ -27,6 +27,7 @@ use Adshares\Adserver\Models\Traits\AutomateMutators;
 use Adshares\Supply\Application\Dto\Info;
 use Adshares\Supply\Domain\ValueObject\HostStatus;
 use Adshares\Supply\Domain\ValueObject\Status;
+use DateTimeImmutable;
 use DateTimeInterface;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
@@ -44,6 +45,7 @@ use Illuminate\Support\Carbon;
  * @property Carbon|null deleted_at
  * @property Carbon last_broadcast
  * @property Carbon|null last_synchronization
+ * @property Carbon|null last_synchronization_attempt
  * @property int failed_connection
  * @property Info info
  * @property string info_url
@@ -58,6 +60,7 @@ class NetworkHost extends Model
     use SoftDeletes;
 
     private const DATETIME_FORMAT = 'Y-m-d H:i:s';
+    private const MAXIMAL_PERIOD_FOR_SYNCHRONIZATION_RETRY_HOURS = 256;
     private const MESSAGE_WHILE_EXCLUDED = 'Server is not on a whitelist';
 
     protected $fillable = [
@@ -76,6 +79,7 @@ class NetworkHost extends Model
     protected $dates = [
         'last_broadcast',
         'last_synchronization',
+        'last_synchronization_attempt',
     ];
 
     public static function fetchByAddress(string $address): ?self
@@ -181,9 +185,29 @@ class NetworkHost extends Model
         return $query->get();
     }
 
+    public static function fetchUnreachableHostsForImportingInventory(array $whitelist = []): Collection
+    {
+        $query = (new self())->where('status', HostStatus::Unreachable);
+        if (!empty($whitelist)) {
+            $query->whereIn('address', $whitelist);
+        }
+
+        return $query->get()->filter(function ($networkHost) {
+            /** @var self $networkHost */
+            $hours = 2 ** max(0, $networkHost->failed_connection - config('app.inventory_failed_connection_limit'));
+            return $hours <= self::MAXIMAL_PERIOD_FOR_SYNCHRONIZATION_RETRY_HOURS &&
+                (
+                    null === $networkHost->last_synchronization_attempt ||
+                    (new DateTimeImmutable(sprintf('-%d hours', $hours)) > $networkHost->last_synchronization_attempt)
+                );
+        });
+    }
+
     public function connectionSuccessful(): void
     {
-        $this->last_synchronization = new Carbon();
+        $now = new Carbon();
+        $this->last_synchronization = $now;
+        $this->last_synchronization_attempt = $now;
         $this->failed_connection = 0;
         $this->status = HostStatus::Operational;
         $this->update();
@@ -191,6 +215,7 @@ class NetworkHost extends Model
 
     public function connectionFailed(): void
     {
+        $this->last_synchronization_attempt = new Carbon();
         ++$this->failed_connection;
         if ($this->failed_connection >= config('app.inventory_failed_connection_limit')) {
             $this->status = HostStatus::Unreachable;