+++ tests/app/Models/NetworkHostTest.php
@@ -29,6 +29,7 @@ use Adshares\Adserver\Utilities\DatabaseConfigReader;
 use Adshares\Supply\Domain\ValueObject\HostStatus;
 use Adshares\Supply\Domain\ValueObject\Status;
 use DateTimeImmutable;
+use Illuminate\Database\Eloquent\Factories\Sequence;
 
 class NetworkHostTest extends TestCase
 {
@@ -132,5 +133,27 @@ class NetworkHostTest extends TestCase
         self::assertEquals(HostStatus::Excluded, $host3->refresh()->status);
         self::assertEquals(HostStatus::Excluded, $host4->refresh()->status);
         self::assertEquals(HostStatus::Failure, $host5->refresh()->status);
+    }
+
+    public function testFetchUnreachableHostsForImportingInventory(): void
+    {
+        NetworkHost::factory()->count(3)
+            ->state(
+                new Sequence(
+                    ['address' => '0001-00000001-8B4E', 'failed_connection' => 10],
+                    ['address' => '0001-00000002-BB2D', 'failed_connection' => 20],
+                    ['address' => '0001-00000003-AB0C', 'failed_connection' => 10],
+                )
+            )
+            ->create([
+                'last_synchronization' => new DateTimeImmutable('-2 days'),
+                'last_synchronization_attempt' => new DateTimeImmutable('-70 minutes'),
+                'status' => HostStatus::Unreachable,
+            ]);
+
+        $hosts = NetworkHost::fetchUnreachableHostsForImportingInventory(['0001-00000001-8B4E', '0001-00000002-BB2D']);
+
+        self::assertCount(1, $hosts);
+        self::assertEquals('0001-00000001-8B4E', $hosts->first()->address);
     }
 }